created api module to handle api requests
I am not sure if this approach is just adding unneccesary layers,
but this is what I am trying at the moment. api module is responsible
for things like specifying http method types and endpoint addresses,
stringifying parameters etc. it returns the response back to the app
module which would take care of app specific duties. we'll see...